+++ src/main/kotlin/kr/flab/f5/f5template/application/ProductController.kt
@@ -1,0 +1,22 @@
+package kr.flab.f5.f5template.application
+
+import io.swagger.annotations.Api
+import org.springframework.web.bind.annotation.PatchMapping
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@Api(tags = ["Product"])
+@RestController
+@RequestMapping("/products")
+class ProductController(
+    private val productService: ProductService
+) {
+
+    @PatchMapping("/{id}/stock")
+    fun decreaseStock(
+        @PathVariable id: Long
+    ) {
+        productService.decreaseStock(id)
+    }
+}
