--- src/main/kotlin/kr/flab/f5/f5template/controller/request/DecreaseStockRequest.kt
@@ -1,8 +1,0 @@
-package kr.flab.f5.f5template.controller.request
-
-import java.math.BigInteger
-
-data class DecreaseStockRequest(
-    val orderNo: BigInteger,
-    val quantity: Int
-)
