--- src/main/kotlin/kr/flab/f5/f5template/controller/request/SetProductRequest.kt
@@ -1,7 +1,0 @@
-package kr.flab.f5.f5template.controller.request
-
-data class SetProductRequest(
-    val name: String,
-    val price: Long,
-    val stock: Long,
-)
