--- src/main/kotlin/kr/flab/f5/f5template/exception/ApiException.kt
@@ -1,8 +1,0 @@
-package kr.flab.f5.f5template.exception
-import org.springframework.http.HttpStatus
-
-class ApiException(
-    val errorMessage: String,
-    val errorType: ErrorType,
-    val httpStatus: HttpStatus
-) : RuntimeException(errorMessage)
