+++ src/main/kotlin/kr/flab/f5/f5template/mysql/jpa/entity/Product.kt
@@ -1,10 +1,7 @@
 package kr.flab.f5.f5template.mysql.jpa.entity
 
-import kr.flab.f5.f5template.exception.ApiException
-import kr.flab.f5.f5template.exception.ErrorType
 import org.hibernate.annotations.CreationTimestamp
 import org.hibernate.annotations.UpdateTimestamp
-import org.springframework.http.HttpStatus
 import java.time.Instant
 import javax.persistence.Column
 import javax.persistence.Entity
@@ -23,13 +20,6 @@ class Product(
     createdAt: Instant = Instant.now(),
     updatedAt: Instant = Instant.now(),
 ) {
-
-    init {
-        if (name.isEmpty()) {
-            throw ApiException("상품명을 입력하지 않았습니다.", ErrorType.INVALID_PARAMETER, HttpStatus.BAD_REQUEST)
-        }
-    }
-
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     var id: Long = id
@@ -65,11 +55,5 @@ class Product(
             throw IllegalArgumentException("Not enough stock for product $id")
         }
         stock -= amount
-    }
-
-    fun reviseProduct(name: String, price: Long, stock: Long) {
-        this.name = name
-        this.price = price
-        this.stock = stock
     }
 }
